Make the module load the WASM module from the same location
The earlier behaviour would have the importing resolve the location to
be relative to the script host document, meaning if you import this
module by "http://example.com/myapp/index.html" this module would
attempt to load the WASM module at "http://example.com/myapp/MD5.wasm",
regardless of where this module ("MD5.js") was loaded from itself.

This commit changes the behaviour so that this module will always look
for "MD5.wasm" "beside" itself -- in the same URL "folder".

This better implements intended coupling between the two -- if you use a
copy of "MD5.js" at "/", for example, then that's where "MD5.wasm" will
be expected to be found as well, regardless of the location of the HTML
document that first (modules are only executed once) caused the importing
of "MD5.js".