Track a particular WASM module implementation
After some consideration, I decided to link to this repository another one which provides the WASM module the `MD5.js` depends on. This will make it easier to use the JS module, and although the tighter coupling is not without drawbacks of its own, the change is done in hopes that effective usafe of what this repository offers, is made easier.